+++ instater/tasks/copy.py
@@ -50,15 +50,34 @@ class Copy(Task):
         self.validate = validate
 
     def _update_metadata(self, file: Path, context: Context) -> bool:
-        return util.update_file_metadata(file, self.owner, self.group, self.mode, context.dry_run)
+        return util.update_file_metadata(file, self.owner, self.group, self.mode, context)
 
     def _validate(self, path: Path):
         if not self.validate:
             return
 
-        validate_command = shlex.split(self.validate % path)
         # util.shell will raise an error on exit codes > 0
-        util.shell(validate_command)
+        util.shell(shlex.split(self.validate % path))
+
+    def _read_for_diff(self, path: Path):
+        if not path.exists():
+            return ""
+
+        try:
+            return _read(path).decode("utf-8")
+        except UnicodeDecodeError:
+            return None
+
+    def _explain_diff(self, src: Union[Path, str], dest: Path, context: Context, src_file=None):
+        # extra condition so we only read the file content if necessary
+        if context.explain:
+            src_content = self._read_for_diff(src) if isinstance(src, Path) else src
+            dest_content = self._read_for_diff(dest)
+
+            if src_content is None or dest_content is None:
+                context.explain_change(f"Binary files differ: {src} and {dest}")
+            else:
+                context.explain_change_diff(dest_content, src_content, str(dest), src_file or str(src))
 
     def _update_file_direct(self, src: Path, dest: Path, context: Context) -> bool:
         updated = False
@@ -66,11 +85,16 @@ class Copy(Task):
         self._validate(src)
 
         if not dest.exists():
+            context.explain_change(f"Destination file does not exist: {dest}")
+            self._explain_diff(src, dest, context)
             if not context.dry_run:
                 dest.parent.mkdir(parents=True, exist_ok=True)
                 shutil.copy(src, dest)
+
             updated = True
         elif not src.samefile(dest) and _read(src) != _read(dest):
+            context.explain_change(f"Source file ({src}) differs from destination file ({dest})")
+            self._explain_diff(src, dest, context)
             if not context.dry_run:
                 shutil.copy(src, dest)
             updated = True
@@ -91,12 +115,14 @@ class Copy(Task):
                 self._validate(Path(temp_file.name))
 
         if not dest.exists():
+            self._explain_diff(content, dest, context, src_file="Template")
             if not context.dry_run:
                 dest.parent.mkdir(parents=True, exist_ok=True)
                 with dest.open("w") as f:
                     f.write(content)
             updated = True
         elif content.encode("utf-8") != _read(dest):
+            self._explain_diff(content, dest, context, src_file="Template")
             if not context.dry_run:
                 with dest.open("w") as f:
                     f.write(content)
@@ -113,9 +139,14 @@ class Copy(Task):
 
     def _update_file(self, src: Path, dest: Path, context: Context):
         if self.is_template:
-            return self._update_file_template(src, dest, context)
+            updated = self._update_file_template(src, dest, context)
         else:
-            return self._update_file_direct(src, dest, context)
+            updated = self._update_file_direct(src, dest, context)
+
+        if not updated:
+            context.explain_skip(f"File {dest} already has the correct content and metadata")
+
+        return updated
 
     def _update_dir(self, context: Context) -> bool:
         updated = False
@@ -156,7 +187,10 @@ class Copy(Task):
                 if self.is_template:
                     content_str = context.jinja_string(content_str)
 
-                return self._update_file_content(content_str, dest, context)
+                updated = self._update_file_content(content_str, dest, context)
+                if not updated:
+                    context.explain_skip(f"File {dest} already has the correct content and metadata")
+                return updated
         else:
             if dest.exists() and not dest.is_dir():
                 raise InstaterError(f"Destination is a file, expected directory: {dest}")