+++ instater/tasks/git.py
@@ -30,9 +30,34 @@ class Git(Task):
         result = util.shell(["git", "config", "--get", "remote.origin.url"], directory=self.dest, become=self.become)
         return result.stdout
 
-    def _should_pull(self) -> bool:
-        result = util.shell(["git", "fetch", "--dry-run", self.tags_flag], directory=self.dest, become=self.become)
-        return result.stdout != "" or result.stderr != ""
+    def _should_pull(self, context: Context) -> bool:
+        # fetch_result captures when remote has changed relative to local
+        fetch_result = util.shell(
+            ["git", "fetch", "--dry-run", self.tags_flag],
+            directory=self.dest,
+            become=self.become,
+        )
+
+        if fetch_result.stdout != "":
+            context.explain_change(f"Local git repository {self.dest} is not up to date: {fetch_result.stdout}")
+            return True
+
+        # log_result captures when the remote has not changed, but
+        # a local branch needs to be fast forwarded (e.g. if the
+        # local repository has been manually reset to a previous
+        # commit, log_result will indicate that a pull should occur)
+        log_result = util.shell(
+            ["git", "log", "-1", "--format=oneline", "@..@{push}"],
+            directory=self.dest,
+            become=self.become,
+        )
+        if log_result.stdout != "":
+            commits = "\n".join(f"  - {line}" for line in log_result.stdout.splitlines())
+            commits_str = f"[white]{commits}[/white]"
+            context.explain_change(f"Local git repository {self.dest} is not up to date. New commits:\n{commits_str}")
+            return True
+
+        return False
 
     def _pull(self):
         branch = util.shell(["git", "branch", "--show-current"], directory=self.dest, become=self.become).stdout
@@ -40,6 +65,7 @@ class Git(Task):
 
     def run_action(self, context: Context):
         if not self.dest.exists():
+            context.explain_change("Git repository has not yet been cloned")
             if not context.dry_run:
                 self._clone()
             return True
@@ -50,9 +76,10 @@ class Git(Task):
         if self._get_remote() != self.repo:
             raise InstaterError(f"Git remote does not match current local git repo: {self.dest}")
 
-        if self._should_pull():
+        if self._should_pull(context):
             if not context.dry_run:
                 self._pull()
             return True
-
-        return False
+        else:
+            context.explain_skip(f"Local git repository {self.dest} is already up to date")
+            return False