+++ instater/tasks/group.py
@@ -15,8 +15,10 @@ class Group(Task):
 
     def run_action(self, context: Context) -> bool:
         if self._group_exists():
+            context.explain_skip(f"Group '{self.group}' already exists")
             return False
 
+        context.explain_change(f"Group '{self.group}' does not yet exist")
         if not context.dry_run:
             util.shell(["groupadd", self.group])